refactor: Rename main demo folder.
