+++ libopenage/main/demo/aicontroller.cpp
