+++ libopenage/main/demo/presenter.cpp
