+++ libopenage/main/demo/presenter.h
