+++ libopenage/main/tests.cpp
@@ -2,8 +2,8 @@
 
 #include "tests.h"
 
-#include "tests/pong.h"
-#include "tests/presenter.h"
+#include "demo/pong.h"
+#include "demo/presenter.h"
 
 
 namespace openage::main::tests {
@@ -11,9 +11,9 @@ namespace openage::main::tests {
 
 void engine_demo(int demo_id, const util::Path &path) {
 	switch (demo_id) {
-	case 0:
+	case 0: {
 		pong::main(path);
-		break;
+	} break;
 
 	case 1: {
 		presenter::demo(path);
